Remove some obsolete code

+++ hmtk/seismicity/max_magnitude/cumulative_moment_release.py
@@ -127,7 +127,7 @@ class CumulativeMoment(BaseMaximumMagnitude):
         # Calculate seismic moment
         m_o = 10. ** (9.05 + 1.5 * mag)
         year_range = np.arange(np.min(year), np.max(year) + 1, 1)
-        nyr = np.float(np.shape(year_range)[0])
+        nyr = np.shape(year_range)[0]
         morate = np.zeros(nyr, dtype=float)
         # Get moment release per year
         for loc, tyr in enumerate(year_range):
